+++ packages/visx-responsive/package.json
@@ -28,10 +28,12 @@
   },
   "homepage": "https://github.com/airbnb/visx#readme",
   "dependencies": {
+    "@juggle/resize-observer": "^3.2.0",
     "@types/lodash": "^4.14.146",
     "@types/react": "*",
     "lodash": "^4.17.10",
     "prop-types": "^15.6.1",
+    "react-use-measure": "^2.0.3",
     "resize-observer-polyfill": "1.5.1"
   },
   "peerDependencies": {
