+++ packages/visx-responsive/src/enhancers/withParentSize.tsx
@@ -1,6 +1,7 @@
-import React, { useRef } from 'react';
+import React from 'react';
+import useMeasure from 'react-use-measure';
+import { ResizeObserver } from '@juggle/resize-observer';
 import { WithSizeProps, WithSizeProvidedProps } from '../hooks/useSize';
-import { useParentSize } from '../hooks/useParentSize';
 
 const CONTAINER_STYLES = { width: '100%', height: '100%' };
 
@@ -8,13 +9,10 @@ export default function withParentSize<BaseComponentProps extends WithSizeProps 
   BaseComponent: React.ComponentType<BaseComponentProps & WithSizeProvidedProps>,
 ) {
   return (props: BaseComponentProps & WithSizeProvidedProps) => {
-    const divRef = useRef(null);
-    const [isParentSize, parentWidth, parentHeight] = useParentSize(props, divRef);
+    const [ref, bounds] = useMeasure({ scroll: true, polyfill: ResizeObserver });
     return (
-      <div style={CONTAINER_STYLES} ref={divRef}>
-        {isParentSize && (
-          <BaseComponent parentWidth={parentWidth} parentHeight={parentHeight} {...props} />
-        )}
+      <div style={CONTAINER_STYLES} ref={ref}>
+        <BaseComponent parentWidth={bounds.width} parentHeight={bounds.height} {...props} />
       </div>
     );
   };
