+++ packages/visx-responsive/src/enhancers/withScreenSize.tsx
@@ -1,16 +1,13 @@
 import React from 'react';
-import { WithSizeProps, WithSizeProvidedProps } from '../hooks/useSize';
-import { useScreenSize } from '../hooks/useScreenSize';
+import { useScreenSize, WithSizeProps, WithSizeProvidedProps } from '../hooks/useScreenSize';
 
-export { WithSizeProvidedProps } from '../hooks/useSize';
+export { WithSizeProvidedProps } from '../hooks/useScreenSize';
 
 export default function withScreenSize<BaseComponentProps extends WithSizeProps = {}>(
   BaseComponent: React.ComponentType<BaseComponentProps>,
 ) {
   return (props: BaseComponentProps & WithSizeProvidedProps) => {
-    const [isScreenSize, screenWidth, screenHeight] = useScreenSize(props);
-    return isScreenSize ? null : (
-      <BaseComponent screenWidth={screenWidth} screenHeight={screenHeight} {...props} />
-    );
+    const [screenWidth, screenHight] = useScreenSize(props);
+    return <BaseComponent screenWidth={screenWidth} screenHeight={screenHight} {...props} />;
   };
 }
